Extract `uniqueFeed` factory helper into a shared scope
