+++ EssentialApp/EssentialAppTests/Helpers/SharedTestHelpers.swift
@@ -1,0 +1,25 @@
+//
+//  SharedTestHelpers.swift
+//  EssentialAppTests
+//
+//  Created by Macbook Pro on 25/9/23.
+//
+
+import Foundation
+import EssentialFeed
+
+func anyNSError() -> NSError {
+    return NSError(domain: "any error", code: 0)
+}
+
+func anyData() -> Data {
+    return Data("any data".utf8)
+}
+
+func anyURL() -> URL {
+    return URL(string: "http://a-url.com")!
+}
+
+func uniqueFeed() -> [FeedImage] {
+    return [FeedImage(id: UUID(), description: "any", location: "any", url: URL(string: "http://any-url.com")!)]
+}
